New main eval page. WIP
