+++ app/web_ui/src/lib/types.ts
@@ -31,3 +31,4 @@ export type EvalRunResult = components["schemas"]["EvalRunResult"]
 export type EvalConfigCompareSummary =
   components["schemas"]["EvalConfigCompareSummary"]
 export type EvalRun = components["schemas"]["EvalRun"]
+export type EvalProgress = components["schemas"]["EvalProgress"]
