+++ app/web_ui/src/routes/(app)/evals/[project_id]/[task_id]/[eval_id]/+page.svelte
@@ -1,35 +1,13 @@
 <script lang="ts">
   import AppPage from "../../../../app_page.svelte"
   import type { Eval } from "$lib/types"
-  import { client, base_url } from "$lib/api_client"
+  import { client } from "$lib/api_client"
   import { KilnError, createKilnError } from "$lib/utils/error_handlers"
   import { onMount, tick } from "svelte"
   import { page } from "$app/stores"
-  import FormElement from "$lib/utils/form_element.svelte"
-  import type {
-    EvalConfig,
-    ProviderModels,
-    TaskRunConfig,
-    EvalResultSummary,
-  } from "$lib/types"
-  import { goto } from "$app/navigation"
-  import {
-    model_info,
-    load_model_info,
-    model_name,
-    provider_name_from_id,
-    prompt_name_from_id,
-    load_available_prompts,
-    load_available_models,
-  } from "$lib/stores"
-  import Dialog from "$lib/ui/dialog.svelte"
-  import AvailableModelsDropdown from "../../../../run/available_models_dropdown.svelte"
-  import PromptTypeSelector from "../../../../run/prompt_type_selector.svelte"
+  import type { EvalProgress } from "$lib/types"
+
   import Warning from "$lib/ui/warning.svelte"
-  import { string_to_json_key } from "$lib/utils/json_schema_editor/json_schema_templates"
-  import RunEval from "./run_eval.svelte"
-  import { eval_config_to_ui_name } from "$lib/utils/formatters"
-  import OutputTypeTablePreview from "./output_type_table_preview.svelte"
   import EditDialog from "$lib/ui/edit_dialog.svelte"
 
   $: project_id = $page.params.project_id
@@ -40,37 +18,18 @@
   let eval_error: KilnError | null = null
   let eval_loading = true
 
-  let eval_configs: EvalConfig[] | null = null
-  let eval_configs_error: KilnError | null = null
-  let eval_configs_loading = true
-  let current_eval_config_id: string | null = null
+  let eval_progress_loading = true
+  let eval_progress: EvalProgress | null = null
+  let eval_progress_error: KilnError | null = null
 
-  let task_run_configs: TaskRunConfig[] | null = null
-  let task_run_configs_error: KilnError | null = null
-  let task_run_configs_loading = true
-
-  let score_summary: EvalResultSummary | null = null
-  let score_summary_error: KilnError | null = null
-
-  // Note: not including score_summary_error, because it's not a critical error we should block the UI for
-  $: loading = eval_loading || eval_configs_loading || task_run_configs_loading
-  $: error = eval_error || eval_configs_error || task_run_configs_error
+  $: loading = eval_loading || eval_progress_loading
+  $: error = eval_error || eval_progress_error
 
   onMount(async () => {
     // Wait for page params to load
     await tick()
-    // Wait for these 3 to load, as they are needed for better labels. Usually already cached and instant.
-    await Promise.all([
-      load_model_info(),
-      load_available_prompts(),
-      load_available_models(),
-    ])
-    // Get the eval first (want it to set the current config id before the other two load)
-    await get_eval()
-    // These two can be parallel
-    await Promise.all([get_eval_configs(), get_task_run_configs()])
-    // This needs the selected eval config id, set from above requests
-    get_score_summary()
+    // Load data in parallel
+    await Promise.all([get_eval(), get_eval_progress()])
   })
 
   async function get_eval() {
@@ -92,11 +51,6 @@
         throw error
       }
       evaluator = data
-      // Set the selected eval config: prefer query params, then eval's default, then first eval config (set below in load_eval_configs)
-      current_eval_config_id =
-        $page.url.searchParams.get("selected_eval_config") ||
-        evaluator.current_config_id ||
-        null
     } catch (error) {
       eval_error = createKilnError(error)
     } finally {
@@ -104,11 +58,13 @@
     }
   }
 
-  async function get_eval_configs() {
+  async function get_eval_progress() {
+    eval_progress = null
+    eval_progress_loading = true
     try {
-      eval_configs_loading = true
+      eval_progress = null
       const { data, error } = await client.GET(
-        "/api/projects/{project_id}/tasks/{task_id}/eval/{eval_id}/eval_configs",
+        "/api/projects/{project_id}/tasks/{task_id}/eval/{eval_id}/progress",
         {
           params: {
             path: {
@@ -122,89 +78,11 @@
       if (error) {
         throw error
       }
-      eval_configs = data
-      // Fallback to first eval config if no current eval config id is set from load_eval()
-      if (
-        !current_eval_config_id &&
-        eval_configs.length > 0 &&
-        eval_configs[0].id
-      ) {
-        current_eval_config_id = eval_configs[0].id
-      }
+      eval_progress = data
     } catch (error) {
-      eval_configs_error = createKilnError(error)
+      eval_progress_error = createKilnError(error)
     } finally {
-      eval_configs_loading = false
-    }
-  }
-
-  async function get_task_run_configs() {
-    try {
-      task_run_configs_loading = true
-      const { data, error } = await client.GET(
-        "/api/projects/{project_id}/tasks/{task_id}/task_run_configs",
-        {
-          params: {
-            path: {
-              project_id,
-              task_id,
-            },
-          },
-        },
-      )
-      if (error) {
-        throw error
-      }
-      task_run_configs = data
-    } catch (error) {
-      task_run_configs_error = createKilnError(error)
-    } finally {
-      task_run_configs_loading = false
-    }
-  }
-
-  async function get_score_summary() {
-    score_summary = null
-    if (!current_eval_config_id) {
-      score_summary_error = new KilnError("No evaluation method selected", null)
-      return
-    }
-    try {
-      score_summary = null
-      const { data, error } = await client.GET(
-        "/api/projects/{project_id}/tasks/{task_id}/eval/{eval_id}/eval_config/{eval_config_id}/score_summary",
-        {
-          params: {
-            path: {
-              project_id,
-              task_id,
-              eval_id,
-              eval_config_id: current_eval_config_id,
-            },
-          },
-        },
-      )
-      if (error) {
-        throw error
-      }
-      score_summary = data
-    } catch (error) {
-      score_summary_error = createKilnError(error)
-    }
-  }
-
-  // Watches the current eval config id, performing actions based on it
-  $: watch_selected_eval_config(current_eval_config_id)
-  function watch_selected_eval_config(selected_id: string | null) {
-    if (selected_id === "add_config") {
-      // if it's the "add_config" special value, navigate to the create eval config page
-      goto(`/evals/${project_id}/${task_id}/${eval_id}/create_eval_config`)
-      return
-    }
-    // If the selected id is not null, then get the score summary
-    score_summary = null
-    if (selected_id) {
-      get_score_summary()
+      eval_progress_loading = false
     }
   }
 
@@ -213,21 +91,9 @@
     value: string
   }
 
-  // A dropdown name for the eval config that is human readable and helpful
-  // Combine's it's name with it's properties
-  function get_eval_config_name(
-    eval_config: EvalConfig,
-    model_info: ProviderModels | null,
-  ): string {
-    let parts = []
-    parts.push(eval_config_to_ui_name(eval_config.config_type))
-    parts.push(model_name(eval_config.model_name, model_info))
-    return eval_config.name + " — " + parts.join(", ")
-  }
-
   function get_eval_properties(
     evaluator: Eval,
-    score_summary: EvalResultSummary | null,
+    eval_progress: EvalProgress | null,
   ): UiProperty[] {
     const properties: UiProperty[] = []
 
@@ -256,8 +122,8 @@
       })
     }
     let eval_set_size = ""
-    if (score_summary) {
-      eval_set_size = " (" + score_summary.dataset_size + " items)"
+    if (eval_progress) {
+      eval_set_size = " (" + eval_progress.dataset_size + " items)"
     }
     properties.push({
       name: "Eval Dataset",
@@ -268,137 +134,6 @@
       value: evaluator.eval_configs_filter_id,
     })
     return properties
-  }
-
-  $: current_eval_config = eval_configs?.find(
-    (config) => config.id === current_eval_config_id,
-  )
-
-  function get_eval_config_properties(
-    eval_config_id: string | null,
-    model_info: ProviderModels | null,
-  ): UiProperty[] {
-    const eval_config = eval_configs?.find(
-      (config) => config.id === eval_config_id,
-    )
-    if (!eval_config) {
-      return [
-        {
-          name: "No Config Selected",
-          value: "Select a config from dropdown above",
-        },
-      ]
-    }
-
-    const properties: UiProperty[] = []
-
-    properties.push({
-      name: "Algorithm",
-      value: eval_config_to_ui_name(eval_config.config_type),
-    })
-    properties.push({
-      name: "Eval Model",
-      value: model_name(eval_config.model_name, model_info),
-    })
-    properties.push({
-      name: "Model Provider",
-      value: provider_name_from_id(eval_config.model_provider),
-    })
-    const task_description = eval_config.properties["task_description"]
-    if (task_description) {
-      properties.push({
-        name: "Task Description",
-        value: task_description,
-      })
-    }
-    return properties
-  }
-
-  function get_eval_config_select_options(
-    configs: EvalConfig[] | null,
-  ): [string, [unknown, string][]][] {
-    const configs_options: [string, string][] = []
-    for (const c of configs || []) {
-      if (c.id) {
-        configs_options.push([c.id, get_eval_config_name(c, $model_info)])
-      }
-    }
-
-    const results: [string, [unknown, string][]][] = []
-    if (configs_options.length > 0) {
-      results.push(["Select Eval Method", configs_options])
-    }
-    results.push(["Manage Eval Methods", [["add_config", "Add Eval Method"]]])
-    return results
-  }
-
-  let eval_state:
-    | "not_started"
-    | "running"
-    | "complete"
-    | "complete_with_errors" = "not_started"
-  $: run_eval_url = `${base_url}/api/projects/${project_id}/tasks/${task_id}/eval/${eval_id}/eval_config/${current_eval_config_id}/run_task_run_eval?all_run_configs=true`
-
-  let task_run_config_model_name = ""
-  let task_run_config_provider_name = ""
-  let task_run_config_prompt_method = "simple_prompt_builder"
-  let task_run_config_long_prompt_name_provider = ""
-
-  let add_task_config_dialog: Dialog | null = null
-  let add_task_config_error: KilnError | null = null
-  async function add_task_config(): Promise<boolean> {
-    if (
-      !task_run_config_model_name ||
-      !task_run_config_provider_name ||
-      !task_run_config_prompt_method
-    ) {
-      add_task_config_error = new KilnError("Missing required fields", null)
-      return false
-    }
-
-    try {
-      const { error } = await client.POST(
-        "/api/projects/{project_id}/tasks/{task_id}/task_run_config",
-        {
-          params: {
-            path: {
-              project_id,
-              task_id,
-            },
-          },
-          body: {
-            model_name: task_run_config_model_name,
-            // @ts-expect-error not checking types here, server will check them
-            model_provider_name: task_run_config_provider_name,
-            prompt_id: task_run_config_prompt_method,
-          },
-        },
-      )
-      if (error) {
-        throw error
-      }
-      // Load the updated list of task run configs after success
-      get_task_run_configs()
-    } catch (error) {
-      add_task_config_error = createKilnError(error)
-      return false
-    }
-    return true
-  }
-
-  function show_incomplete_warning(
-    score_summary: EvalResultSummary | null,
-  ): boolean {
-    if (!score_summary?.run_config_percent_complete) {
-      return false
-    }
-
-    const values = Object.values(score_summary.run_config_percent_complete)
-    const minComplete =
-      values.length > 0
-        ? values.reduce((min, val) => Math.min(min, val), 1.0)
-        : 1.0
-    return minComplete < 1.0
   }
 
   $: has_default_eval_config = evaluator && evaluator.current_config_id
@@ -439,288 +174,45 @@
   {:else if evaluator}
     <div class="flex flex-col xl:flex-row gap-8 xl:gap-16 mb-8">
       <div class="grow basis-1/2">
+        <ul class="steps steps-vertical ml-4">
+          <li class="step step-primary" data-content="✓">Define Goals</li>
+          <li class="step step-primary" data-content="✓">Create Eval Data</li>
+          <li class="step step-primary" data-content="✓">Rate Eval Data</li>
+          <li class="step" data-content="">Find Ideal Eval Method</li>
+          <li class="step" data-content="">
+            <a
+              class="link"
+              href={`/evals/${project_id}/${task_id}/${eval_id}/compare_run_methods`}
+              >Find Ideal Way to Run Task</a
+            >
+          </li>
+        </ul>
+      </div>
+      <div class="grow basis-1/2">
         <div class="text-xl font-bold mb-4">Evaluator Properties</div>
         <div
           class="grid grid-cols-[auto,1fr] gap-y-2 gap-x-4 text-sm 2xl:text-base"
         >
-          {#each get_eval_properties(evaluator, score_summary) as property}
+          {#each get_eval_properties(evaluator, eval_progress) as property}
             <div class="flex items-center">{property.name}</div>
             <div class="flex items-center text-gray-500 overflow-x-hidden">
               {property.value}
             </div>
           {/each}
         </div>
-        {#if score_summary && score_summary.dataset_size > 0 && score_summary.dataset_size < 25}
+        {#if eval_progress && eval_progress.dataset_size > 0 && eval_progress.dataset_size < 25}
           <div class="mt-4">
             <Warning
-              warning_message={`There are only ${score_summary.dataset_size} item(s) in your eval dataset. This is generally too small to get a good sense of how well your task run methods perform.`}
+              warning_message={`There are only ${eval_progress.dataset_size} item(s) in your eval dataset. This is generally too small to get a good sense of how well your task run methods perform.`}
               warning_color="warning"
               tight={true}
             />
           </div>
         {/if}
       </div>
-      <div class="grow basis-1/2 flex flex-col gap-4">
-        <div>
-          <div class="text-xl font-bold">Evaluation Method</div>
-          <div class="text-sm text-gray-500 mb-2">
-            How the task outputs will be evaluated.
-          </div>
-
-          <FormElement
-            hide_label={true}
-            id="eval_config_select"
-            label="Eval Method"
-            inputType="select"
-            bind:value={current_eval_config_id}
-            select_options_grouped={get_eval_config_select_options(
-              eval_configs,
-            )}
-          />
-          {#if !has_default_eval_config}
-            <Warning
-              warning_message="No default evaluation method selected. We recommend using 'Compare Evaluation Methods' and selecting the best performing method as the default."
-              warning_color="warning"
-              tight={true}
-            />
-          {:else if has_default_eval_config && evaluator.current_config_id != current_eval_config_id}
-            <Warning
-              warning_message="The currently selected evaluation method is not the default. You can change the default in 'Compare Evaluation Methods'."
-              warning_color="warning"
-              tight={true}
-            />
-          {/if}
-        </div>
-        <div
-          class="grid grid-cols-[auto,1fr] gap-y-2 gap-x-4 text-sm 2xl:text-base"
-        >
-          {#each get_eval_config_properties(current_eval_config_id, $model_info) as property}
-            <div class="flex items-center">{property.name}</div>
-            <div class="flex items-center text-gray-500 overflow-x-hidden">
-              {property.value}
-            </div>
-          {/each}
-          <div class="flex items-center">Eval Method Quality</div>
-          <div class="flex items-center text-gray-500 overflow-x-hidden">
-            <a
-              href={`/evals/${project_id}/${task_id}/${eval_id}/eval_configs`}
-              class="link"
-            >
-              Compare and optimize
-            </a>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="mt-16">
-      {#if task_run_configs?.length}
-        <div class="flex flex-col lg:flex-row gap-4 lg:gap-8 mb-6">
-          <div class="grow">
-            <div class="text-xl font-bold">Compare Run Methods</div>
-            <div class="text-xs text-gray-500">
-              Find the best method of running your task comparing various
-              prompts, models, fine-tunes, and more.
-            </div>
-            <div class="text-xs text-gray-500 pt-2">
-              Scores are generated by running the 'run method' on each item of
-              your eval dataset, generating task outputs, then evaluating those
-              outputs with the selected evaluation method{current_eval_config
-                ? ` (${current_eval_config.name})`
-                : ""}.
-            </div>
-            {#if score_summary_error}
-              <div class="text-error text-sm">
-                {score_summary_error.getMessage() ||
-                  "An unknown error occurred fetching scores."}
-              </div>
-            {/if}
-          </div>
-          <div class="shrink-0">
-            {#if eval_state === "not_started"}
-              <button
-                class="btn btn-mid mr-2"
-                on:click={() => {
-                  add_task_config_dialog?.show()
-                }}>Add Run Method</button
-              >
-            {/if}
-            <RunEval
-              bind:eval_state
-              bind:run_url={run_eval_url}
-              on_run_complete={() => {
-                get_score_summary()
-              }}
-            />
-          </div>
-        </div>
-
-        <!-- Warn the user if some evals are incomplete -->
-        {#if show_incomplete_warning(score_summary)}
-          <div class="mt-6 mb-4">
-            <button
-              class="tooltip tooltip-top cursor-pointer"
-              data-tip="Running evals will update any missing dataset items, without re-running complete items. If some evals consistently fail, check the logs for error details."
-            >
-              <Warning
-                warning_message={`Some evals are incomplete and should be excluded from analysis. Click 'Run Eval' to generate missing results.`}
-                tight={true}
-              />
-            </button>
-          </div>
-        {/if}
-
-        <div class="overflow-x-auto rounded-lg border">
-          <table class="table">
-            <thead>
-              <tr>
-                <th>
-                  <div>Run Method</div>
-                  <div class="font-normal">How task output is generated</div>
-                </th>
-                {#each evaluator.output_scores as output_score}
-                  <th class="text-center">
-                    {output_score.name}
-                    <OutputTypeTablePreview
-                      output_score_type={output_score.type}
-                    />
-                  </th>
-                {/each}
-              </tr>
-            </thead>
-            <tbody>
-              {#each task_run_configs || [] as task_run_config}
-                {@const percent_complete =
-                  score_summary?.run_config_percent_complete?.[
-                    "" + task_run_config.id
-                  ]}
-                <tr
-                  class="hover cursor-pointer"
-                  on:click={() => {
-                    goto(
-                      `/evals/${project_id}/${task_id}/${eval_id}/${current_eval_config_id}/${task_run_config.id}/run_result`,
-                    )
-                  }}
-                >
-                  <td>
-                    <div class="font-medium">
-                      {task_run_config.name}
-                    </div>
-                    <div class="text-sm text-gray-500">
-                      {model_name(
-                        task_run_config?.run_config_properties?.model_name,
-                        $model_info,
-                      )}
-                    </div>
-                    <div class="text-sm text-gray-500">
-                      {provider_name_from_id(
-                        task_run_config?.run_config_properties
-                          ?.model_provider_name,
-                      )}
-                    </div>
-                    <div class="text-sm text-gray-500">
-                      Prompt Name:
-                      {task_run_config.prompt?.name ||
-                        prompt_name_from_id(
-                          task_run_config?.run_config_properties?.prompt_id,
-                        )}
-                    </div>
-                    {#if task_run_config?.prompt?.generator_id && task_run_config?.run_config_properties?.prompt_id?.startsWith("task_run_config::")}
-                      <!-- Special description for prompts frozen to the task run config. The name alone isn't that helpful, so we say where it comes from (eg "Basic (Zero Shot")) -->
-                      <div class="text-sm text-gray-500">
-                        Prompt Source: {prompt_name_from_id(
-                          task_run_config?.prompt?.generator_id,
-                        )}
-                      </div>
-                    {/if}
-                    {#if percent_complete}
-                      <div
-                        class="text-sm {percent_complete < 1.0
-                          ? 'text-error'
-                          : 'text-gray-500'}"
-                      >
-                        {(percent_complete * 100.0).toFixed(1)}% complete
-                      </div>
-                    {:else if score_summary}
-                      <!-- We have results, but not for this run config -->
-                      <div class="text-sm text-error">0% complete</div>
-                    {/if}
-                  </td>
-                  {#each evaluator.output_scores as output_score}
-                    {@const score =
-                      score_summary?.results?.["" + task_run_config.id]?.[
-                        string_to_json_key(output_score.name)
-                      ]?.mean_score}
-                    <td class="text-center">
-                      {score != null ? score.toFixed(2) : "unknown"}
-                    </td>
-                  {/each}
-                </tr>
-              {/each}
-            </tbody>
-          </table>
-        </div>
-      {:else}
-        <div class="text-xl font-bold">Compare Run Methods</div>
-        <div class="text-sm text-gray-500">
-          Find the best method of running your task comparing various prompts,
-          models, fine-tunes, and more. Add one or more task run methods to get
-          started.
-        </div>
-
-        <button
-          class="btn min-w-[200px] mt-4 {has_default_eval_config
-            ? 'btn-primary'
-            : ''}"
-          on:click={() => {
-            add_task_config_dialog?.show()
-          }}
-        >
-          Add Run Method
-        </button>
-      {/if}
     </div>
   {/if}
 </AppPage>
-
-<Dialog
-  bind:this={add_task_config_dialog}
-  title="Add a Task Run Method"
-  action_buttons={[
-    {
-      label: "Cancel",
-      isCancel: true,
-    },
-    {
-      label: "Create",
-      isPrimary: true,
-      asyncAction: add_task_config,
-    },
-  ]}
->
-  <h4 class="text-sm text-gray-500">
-    Define a method of running this task (model+prompt).
-  </h4>
-  <h4 class="text-sm text-gray-500 mt-1">
-    Your evaluator can compare multiple run methods to find which one produces
-    the highest scores on your eval dataset.
-  </h4>
-  <div class="flex flex-col gap-2 pt-6">
-    <AvailableModelsDropdown
-      bind:model_name={task_run_config_model_name}
-      bind:provider_name={task_run_config_provider_name}
-      bind:model={task_run_config_long_prompt_name_provider}
-    />
-    <PromptTypeSelector
-      bind:prompt_method={task_run_config_prompt_method}
-      bind:linked_model_selection={task_run_config_long_prompt_name_provider}
-    />
-    {#if add_task_config_error}
-      <div class="text-error text-sm">
-        {add_task_config_error.getMessage() || "An unknown error occurred"}
-      </div>
-    {/if}
-  </div>
-</Dialog>
 
 <EditDialog
   bind:this={edit_dialog}
