+++ app/web_ui/src/routes/(app)/evals/[project_id]/[task_id]/[eval_id]/compare_run_methods/+page.ts
@@ -1,0 +1,1 @@
+export const prerender = false
